+++ DatabaseInterpreter/DatabaseInterpreter.Model/DataType/DataTypeInfo.cs
@@ -4,8 +4,8 @@
     {
         public string DataType { get; set; }       
         public long? MaxLength { get; set; }
-        public int? Precision { get; set; }
-        public int? Scale { get; set; }
+        public long? Precision { get; set; }
+        public long? Scale { get; set; }
         public bool IsIdentity { get; set; }
         public string Args { get; set; }
